requirements.txt
requests~=2.32.4
bs4~=0.0.2
beautifulsoup4~=4.13.4
pandas~=2.3.2
PyYAML~=6.0.2
matplotlib~=3.10.6
joblib~=1.5.2
scikit-learn~=1.7.2
fastapi~=0.116.1
numpy~=2.3.2
pydantic~=2.11.9
streamlit~=1.49.1